add status update.

+++ src/Pedidos.Domain/Models/Item.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int PedidoId { get; set; }
         public string Descricao { get; set; }
-        public double PrecoUnitario { get; set; }
+        public decimal PrecoUnitario { get; set; }
         public int Quantidade { get; set; }
     }
 }
